moving the register controller to the user routes
